Moved model tests into model subdirectory
